third_party/WebKit/LayoutTests/virtual/color_space/fast/canvas/color-space/canvas-colorManaged-toBlob-toDataURL.html

<!DOCTYPE HTML>
<script src="../../../../../resources/testharness.js"></script>
<script src="../../../../../resources/testharnessreport.js"></script>
<script>

function testPixels(actualPixels, expectedPixels, pixelFormat)
{
    var tolerance = 0;
    if (pixelFormat == "float16")
        tolerance = 0.01;
    assert_array_approx_equals(actualPixels, expectedPixels, tolerance);
}


function generateFillStyle(red, green, blue, alpha) {
    return "rgba(" + red + "," + green + "," + blue + "," + alpha + ")";
}

function createTestCanvasAndImageData(testScenario)
{
    var canvas = document.createElement("canvas");
    canvas.width = 2;
    canvas.height = 2;
    var ctx = canvas.getContext('2d', {colorSpace: testScenario.colorSpace,
                                       pixelFormat: testScenario.pixelFormat});
    ctx.fillStyle = generateFillStyle(155, 27, 27, testScenario.alpha);
    ctx.fillRect(0, 0, 1, 1);
    ctx.fillStyle = generateFillStyle(27, 155, 27, testScenario.alpha);
    ctx.fillRect(1, 0, 1, 1);
    ctx.fillStyle = generateFillStyle(27, 27, 155, testScenario.alpha);
    ctx.fillRect(0, 1, 1, 1);
    ctx.fillStyle = generateFillStyle(27, 27, 27, testScenario.alpha);
    ctx.fillRect(1, 1, 1, 1);
    imageData = ctx.getImageData(0, 0, 2, 2).dataUnion;
    var result = {};
    result.canvas = canvas;
    result.imageData = imageData;
    return result;
}

function testScenarioToString(testScenario) {
    return "Color space: " + testScenario.colorSpace +
           ", pixel format: " + testScenario.pixelFormat +
           ", alpha: " + testScenario.alpha +
           ", colorSpaceConversion: " + testScenario.colorSpaceConversion;
}

function runToBlobTest(testScenario) {
    var initVars = createTestCanvasAndImageData(testScenario);
    var srcCanvas = initVars.canvas;
    var expectedPixels = initVars.imageData;
    async_test(t => {
        srcCanvas.toBlob(function(blob) {
            var options = {colorSpaceConversion: testScenario.colorSpaceConversion};
            createImageBitmap(blob, options).then(function(bitmap) {
                var dstCanvas = document.createElement("canvas");
                dstCanvas.width = 2;
                dstCanvas.height = 2;
                var ctx = dstCanvas.getContext('2d',
                        {colorSpace: testScenario.colorSpace,
                         pixelFormat: testScenario.pixelFormat});
                ctx.drawImage(bitmap, 0, 0);
                var actualPixels = ctx.getImageData(0, 0, 2, 2).dataUnion;
                testPixels(actualPixels, expectedPixels, testScenario.pixelFormat);
                t.done();
            });
        }, 'image/png', 1);
   }, "Test if toBlob is color managed: " + testScenarioToString(testScenario));
}

function runToDataURLTest(testScenario) {
    var initVars = createTestCanvasAndImageData(testScenario);
    var srcCanvas = initVars.canvas;
    var expectedPixels = initVars.imageData;

    var image = new Image();
    var t_dataurl = async_test("Test if toDataURL is color managed: " +
        testScenarioToString(testScenario));
    image.onload = t_dataurl.step_func(function() {
        var options = {colorSpaceConversion: testScenario.colorSpaceConversion};
        var imagebitmap;
        var p1 = createImageBitmap(image, options).then(function(bitmap) {
            imagebitmap = bitmap });
        return Promise.all([p1]).then(
            t_dataurl.step_func_done(function() {
                var dstCanvas = document.createElement("canvas");
                dstCanvas.width = 2;
                dstCanvas.height = 2;
                var ctx = dstCanvas.getContext('2d',
                        {colorSpace: testScenario.colorSpace,
                        pixelFormat: testScenario.pixelFormat});
                ctx.drawImage(imagebitmap, 0, 0);
                var actualPixels = ctx.getImageData(0, 0, 2, 2).dataUnion;
                testPixels(actualPixels, expectedPixels, testScenario.pixelFormat);
        }), t_dataurl.step_func_done(function() {
            assert_true(false, "createImageBitmap promise rejected.");
        }));
    });
    image.src = srcCanvas.toDataURL();
}

function runAllTests() {
    var testScenarioSet = [
        {colorSpace: 'srgb', pixelFormat: '8-8-8-8',
            colorSpaceConversion: 'default', alpha: 1},
        {colorSpace: 'srgb', pixelFormat: 'float16',
            colorSpaceConversion: 'linear-rgb', alpha: 1},
        {colorSpace: 'rec2020', pixelFormat: 'float16',
            colorSpaceConversion: 'rec2020', alpha: 1},
        {colorSpace: 'p3', pixelFormat: 'float16',
            colorSpaceConversion: 'p3', alpha: 1},

        {colorSpace: 'srgb', pixelFormat: '8-8-8-8',
            colorSpaceConversion: 'default', alpha: 0.5},
        {colorSpace: 'srgb', pixelFormat: 'float16',
            colorSpaceConversion: 'linear-rgb', alpha: 0.5},
        {colorSpace: 'rec2020', pixelFormat: 'float16',
            colorSpaceConversion: 'rec2020', alpha: 0.5},
        {colorSpace: 'p3', pixelFormat: 'float16',
            colorSpaceConversion: 'p3', alpha: 0.5},
    ];

    for (var i = 0; i < testScenarioSet.length; i++)
        runToBlobTest(testScenarioSet[i]);

    for (var i = 0; i < testScenarioSet.length; i++)
        runToDataURLTest(testScenarioSet[i]);
}

runAllTests();

</script>
